requirements.txt
# This file was autogenerated by uv via the following command:
#    ./scripts/generate_requirements.sh
agno-aws==0.0.1
agno-docker==0.0.1
aiofiles==24.1.0
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.15
botocore==1.38.15
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
curl-cffi==0.11.0
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docstring-parser==0.16
duckduckgo-search==8.0.1
email-validator==2.2.0
exa-py==1.13.1
fastapi==0.115.12
fastapi-cli==0.0.7
feedparser==6.0.11
filelock==3.18.0
frozendict==2.4.6
gitdb==4.0.12
gitpython==3.1.44
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lxml==5.4.0
lxml-html-clean==0.4.2
mako==1.3.10
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
multitasking==0.0.11
narwhals==1.39.0
nest-asyncio==1.6.0
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==2.2.5
openai==1.78.1
packaging==24.2
pandas==2.2.3
peewee==3.18.1
pgvector==0.4.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.5.0
primp==0.15.0
protobuf==6.30.2
psycopg==3.2.9
psycopg-binary==3.2.9
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.4
pydantic-core==2.33.2
pydantic-settings==2.9.1
pydeck==0.9.1
pygments==2.19.1
pypdf==5.5.0
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==14.0.0
rich-toolkit==0.14.6
rpds-py==0.24.0
s3transfer==0.12.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.40
starlette==0.46.2
streamlit==1.45.1
tenacity==9.1.2
tiktoken==0.9.0
tldextract==5.3.0
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
typer==0.15.3
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
yfinance==0.2.61
